Create MEP_Adventurer
Hello all! 

I'm a big fan of LotR and CkII, so I was natural for me to help for this mod as much as I can! :) 
Especially, I was focused on colonization process, and for this I created a new councillor: the Adventurer !

I transfered the action_colonize_province from the Steward to him, and I created the events... and it works!
However, I have two littles problems with this new councillor : 

1) The AI uses him as an Emissary... I mean, the Adventurer counts as an Emissary with Emissary's actions, even when I put "is_adventurer = yes" in job_titles.
2) I created an other action for this councillor called "Spread Local Tensions" (little wink to Grima). The events works with console commands, but in-game the message doesn't pop up and the events are not activated because of this...

If you can watch what I've done to tell me what's wrong, It will be very nice from you :) 

I work on 2.7.1 CKII version for now, and for more events (plot to become an Adventurer,...), I wait to fix these first problems before